Fixes empty control exception when a document type is not saved after adding a cell
Steps to reproduce exception:

- Choose a layout
- Add an element
- Select any DTGE element
- Choose the Cancel button

A layout should then contain an empty cell which holds no data due to the document type instance not being saved. When rendering the grid on a page template, the following change is required as the `docTypeAlias` parameter is empty.

I'm not sure whether the issue is related to the grid or the DTGE but the change proposed appears to fix the issue.