Cancels preview request if redirected
If the DTGE preview result is a redirect, then we cancel the request.
The issue here is that typically a redirect from a controller would be to a full-loading HTML page.
Meaning that a full bodied HTML page would be injected into a DTGE grid cell, causing havoc.
This is a temporary resolution until we implement a more robust DTGE previewer mechanism.