Stopped rendered contents from having click events
The only way to edit the doc type is by clicking on the rendered output. If the render has a link in it, then clicking on it will open the link instead of allowing you to edit the doc type. Adding "prevent-default" prevents the default for when a link is clicked on.