Changed the parameter name to "pageId"
This follows our approach in Stacked Content and gives the intent it for the Page, rather than a "node",
(which becomes confusing in this modern NC/DTGE/StackedContent world)

Added an additional check that the page ID is greater than 0.
A newly created page has an ID of -1, which means that it isn't in the cache or database and would throw an exception.