Removed setting the `InPreviewMode` flag
The change we introduced from #117 to trick Umbraco in thinking that it
was in full preview mode has had some bad side-effects, see issues #123,
#124 and #125 for further details.

This change removes us manually setting the `InPreviewMode` property,
and adds "`dtgePreview`" to the partial-view's `ViewData`, so a
developer can detect whether its rendering in DTGE's preview mode or
not.

I'm unsure how this change will impact #117, there is an extra level of
complexity when attempting cross-package-compatibility, (in this case
Vorto).